trying to fix float

+++ train_distilled_image.py
@@ -98,7 +98,7 @@ class Trainer(object):
         for step_i, (data, label, lr) in enumerate(steps):
             with torch.enable_grad():
                 output = model.forward_with_param(data, w)
-                loss = task_loss(state, output.float(), label)
+                loss = task_loss(state, output.float(), label.float())
             gw, = torch.autograd.grad(loss, w, lr, create_graph=True)
 
             with torch.no_grad():
